+++ PirateLexer/Lexer.cs
@@ -1,4 +1,6 @@
 
+using Common;
+using Common.Enum;
 using PirateLexer.Models;
 
 namespace PirateLexer;
@@ -8,9 +10,11 @@ public class Lexer
     public static string text { get; set; }
     public static char currentChar { get; set; }
     public static Position position { get; set; }
+    public Logger logger { get; set; }
 
-    public Lexer(string FileName, string Text)
+    public Lexer(string FileName, string Text, Logger Logger)
     {
+        logger = Logger;
         fileName = FileName;
         var newText = Text.Replace("\n", "").Replace("\r", "").Replace("    ", "");
         text = newText;
@@ -35,12 +39,14 @@ public class Lexer
     {
         List<Token> tokens = new List<Token> { };
 
+        logger.Log("Starting Lexing text", "Lexer", LogType.INFO);
         while (currentChar != null)
         {
             if (currentChar.Equals('\n'))
             {
                 tokens.Add(new Token(
                     TokenType.ENDOFLINE,
+                    logger,
                     PositionStart: position
                 ));
                 Advance();
@@ -56,26 +62,27 @@ public class Lexer
             }
             if (Globals.DIGITS.Contains(currentChar))
             {
-                tokens.Add(TokenRepository.MakeNumber());
+                tokens.Add(TokenRepository.MakeNumber(logger));
                 continue;
             }
             if (Globals.LETTERS.Contains(currentChar))
             {
-                tokens.Add(TokenRepository.MakeIdentifier());
+                tokens.Add(TokenRepository.MakeIdentifier(logger));
                 continue;
             }
             switch (currentChar)
             {
                 case '"':
-                    tokens.Add(TokenRepository.MakeString());
+                    tokens.Add(TokenRepository.MakeString(logger));
                     continue;
                 case '+':
-                    tokens.Add(TokenRepository.MakePlus());
+                    tokens.Add(TokenRepository.MakePlus(logger));
                     Advance();
                     continue;
                 case '-':
                     tokens.Add(new Token(
                         TokenType.MINUS,
+                        logger,
                         PositionStart: position
                     ));
                     Advance();
@@ -83,17 +90,19 @@ public class Lexer
                 case '*':
                     tokens.Add(new Token(
                         TokenType.MULTIPLY,
+                        logger,
                         PositionStart: position
                     ));
                     Advance();
                     continue;
                 case '/':
-                    tokens.Add(TokenRepository.MakeDivide());
+                    tokens.Add(TokenRepository.MakeDivide(logger));
                     Advance();
                     continue;
                 case '^':
                     tokens.Add(new Token(
                         TokenType.POWER,
+                        logger,
                         PositionStart: position
                     ));
                     Advance();
@@ -101,6 +110,7 @@ public class Lexer
                 case '(':
                     tokens.Add(new Token(
                         TokenType.LEFTPARENTHESES,
+                        logger,
                         PositionStart: position
                     ));
                     Advance();
@@ -108,6 +118,7 @@ public class Lexer
                 case ')':
                     tokens.Add(new Token(
                         TokenType.RIGHTPARENTHESES,
+                        logger,
                         PositionStart: position
                     ));
                     Advance();
@@ -115,6 +126,7 @@ public class Lexer
                 case '{':
                     tokens.Add(new Token(
                         TokenType.LEFTCURLYBRACE,
+                        logger,
                         PositionStart: position
                     ));
                     Advance();
@@ -122,6 +134,7 @@ public class Lexer
                 case '}':
                     tokens.Add(new Token(
                         TokenType.RIGHTCURLYBRACE,
+                        logger,
                         PositionStart: position
                     ));
                     Advance();
@@ -129,6 +142,7 @@ public class Lexer
                 case ',':
                     tokens.Add(new Token(
                         TokenType.COMMA,
+                        logger,
                         PositionStart: position
                     ));
                     Advance();
@@ -136,6 +150,7 @@ public class Lexer
                 case ':':
                     tokens.Add(new Token(
                         TokenType.COLON,
+                        logger,
                         PositionStart: position
                     ));
                     Advance();
@@ -143,6 +158,7 @@ public class Lexer
                 case ';':
                     tokens.Add(new Token(
                         TokenType.SEMICOLON,
+                        logger,
                         PositionStart: position
                     ));
                     Advance();
@@ -150,6 +166,7 @@ public class Lexer
                 case '.':
                     tokens.Add(new Token(
                         TokenType.DOT,
+                        logger,
                         PositionStart: position
                     ));
                     Advance();
@@ -157,6 +174,7 @@ public class Lexer
                 case '$':
                     tokens.Add(new Token(
                         TokenType.DOLLAR,
+                        logger,
                         PositionStart: position
                     ));
                     Advance();
@@ -164,6 +182,7 @@ public class Lexer
                 case '[':
                     tokens.Add(new Token(
                         TokenType.LEFTBRACKET,
+                        logger,
                         PositionStart: position
                     ));
                     Advance();
@@ -171,24 +190,25 @@ public class Lexer
                 case ']':
                     tokens.Add(new Token(
                         TokenType.RIGHTBRACKET,
+                        logger,
                         PositionStart: position
                     ));
                     Advance();
                     continue;
                 case '=':
-                    tokens.Add(TokenRepository.MakeEquals());
+                    tokens.Add(TokenRepository.MakeEquals(logger));
                     Advance();
                     continue;
                 case '<':
-                    tokens.Add(TokenRepository.MakeLessThan());
+                    tokens.Add(TokenRepository.MakeLessThan(logger));
                     Advance();
                     continue;
                 case '>':
-                    tokens.Add(TokenRepository.MakeGreaterThan());
+                    tokens.Add(TokenRepository.MakeGreaterThan(logger));
                     Advance();
                     continue;
                 case '!':
-                    var result = TokenRepository.MakeNotEquals();
+                    var result = TokenRepository.MakeNotEquals(logger);
                     if (result.error != null)
                     {
                         return (null, result.error);
@@ -199,14 +219,11 @@ public class Lexer
                 default:
                     var positionStart = position.Copy();
                     Advance();
+                    logger.Log($"{currentChar} not found in the Lexer", "Lexer", LogType.ERROR);
                     return (null, new Error(positionStart, position, $"'{currentChar}'", "Illegal Character"));
             }
         }
-
-        tokens.Add(new Token(
-            TokenType.ENDOFFILE,
-            PositionStart: position
-        ));
+        logger.Log("Finished Lexing\n", "Lexer", LogType.INFO);
         return (tokens, null);
     }
 }