+++ Shell/Commands/CommandFactory.cs
@@ -1,17 +1,21 @@
+using Common;
+
 namespace Shell.Commands
 {
     public class CommandFactory
     {
-        public static ICommand GetCommand(string commandArgument, string version)
+        public static ICommand GetCommand(string commandArgument, string version, Logger logger)
         {
             switch (commandArgument)
             {
                 case "init":
-                    return new InitCommand(version);
+                    return new InitCommand(version, logger);
                 case "new":
-                    return new NewCommand(version);
+                    return new NewCommand(version, logger);
                 case "run":
-                    return new RunCommand(version);
+                    return new RunCommand(version, logger);
+                case "build":
+                    return new BuildCommand(version, logger);
                 default:
                     return null;
             }