+++ Shell/Commands/InitCommand.cs
@@ -1,22 +1,24 @@
-
-using PirateInterpreter;
-using PirateLexer;
-using PirateParser;
+using Common;
+using Common.Enum;
 
 namespace Shell.Commands
 {
     public class InitCommand : ICommand
     {
         public string version { get; set; }
-        public InitCommand(string Version)
+        public Logger logger { get; set; }
+        public InitCommand(string Version, Logger Logger)
         {
             version = Version;
+            logger = Logger;
         }
         public void Run(string[] arguments)
         {
+            logger.Log("Starting Init Command", this.GetType().Name, LogType.INFO);
             var nameArgument = "main";
             if (arguments.Length == 2) { nameArgument = arguments[1];}
 
+            logger.Log($"Creating {nameArgument} file", this.GetType().Name, LogType.INFO);
             var fileName = nameArgument.Replace(".pirate", "");
             var file = File.CreateText($"./{fileName}.pirate");
             file.Write(String.Join(
@@ -27,6 +29,7 @@ namespace Shell.Commands
                 "}"
             ));
             file.Close();
+            logger.Log($"Created {nameArgument} file", this.GetType().Name, LogType.INFO);
             Console.WriteLine($"\nCreated {fileName}.pirate");
         }
 
@@ -41,6 +44,12 @@ namespace Shell.Commands
                 "\nOptions",
                 "   -h --help       Show command line help."
             ));
+        }
+        public void Error(string message)
+        {
+            Console.ForegroundColor = ConsoleColor.Red;
+            Console.WriteLine($"\n{message}");
+            Console.ForegroundColor = ConsoleColor.White;
         }
     }
 }