+++ Shell/Commands/NewCommand.cs
@@ -1,25 +1,26 @@
-using System.Security.AccessControl;
-
-using PirateInterpreter;
-using PirateLexer;
-using PirateParser;
+using Common;
+using Common.Enum;
 
 namespace Shell.Commands
 {
     public class NewCommand : ICommand
     {
         public string version { get; set; }
-        public NewCommand(string Version)
+        public Logger logger { get; set; }
+        public NewCommand(string Version, Logger Logger)
         {
             version = Version;
+            logger = Logger;
         }
         public void Run(string[] arguments)
         {
+            logger.Log("Starting New Command", this.GetType().Name, LogType.INFO);
             var typeArgument = string.Empty;
             if (arguments.Length >= 2) { typeArgument = arguments[1]; }
 
             if (typeArgument == string.Empty)
             {
+                logger.Log("Argument is empty", this.GetType().Name, LogType.INFO);
                 Console.WriteLine(String.Join(
                     Environment.NewLine,
                     "\nThe \"pirate new [type]\" command creates a new file from a template",
@@ -38,9 +39,12 @@ namespace Shell.Commands
             };
             if (!typeOptions.Contains(typeArgument))
             {
+                logger.Log($"Specified file \"{typeArgument}\" not able to be created", this.GetType().Name, LogType.ERROR);
                 Error($"Specified file \"{typeArgument}\" not able to be created");
                 return;
             }
+
+            logger.Log($"Creating {typeArgument} file", this.GetType().Name, LogType.INFO);
             switch (typeArgument)
             {
                 case "gitignore":
@@ -63,6 +67,7 @@ namespace Shell.Commands
                     var exists = File.Exists($"./{filename}.pirate");
                     if (exists)
                     {
+                        logger.Log($"Specified filename \"{filename}\" already exists", this.GetType().Name, LogType.WARNING);
                         Error($"Specified filename \"{filename}\" already exists");
                         return;
                     }